Misc update

+++ resources/views/livewire/website-product-list-display.blade.php
@@ -10,7 +10,11 @@
 
       <div class="d-flex flex-column justify-content-between h-100 bg-success-rm">
           <div class="d-flex justify-content-center bg-warning-rm">
-            <img class="img-fluid h-25-rm w-100-rm" src="{{ asset('storage/' . $product->image_path) }}" alt="{{ $product->name }}" style="max-height: 150px; {{--max-width: 100px;--}}">
+            @if ($product->image_path)
+              <img class="img-fluid h-25-rm w-100-rm" src="{{ asset('storage/' . $product->image_path) }}" alt="{{ $product->name }}" style="max-height: 150px; {{--max-width: 100px;--}}">
+            @else
+              <i class="fas fa-dice-d6 fa-8x text-muted m-5"></i>
+            @endif
           </div>
 
         <div class="d-flex flex-column justify-content-between flex-grow-1 overflow-auto" style="background-color: #f5f5f5;">
